cleaning code, remove setting function
