+++ src/main/java/business/InvoiceService.java
@@ -6,10 +6,61 @@ import java.util.List;
 import java.util.Date;
 
 /**
- * Enhanced Invoice Service with proper business operations
+ * Enhanced Invoice Service with proper business operations and embedded SearchCriteria
  * Replaces the simple addInvoice class with comprehensive CRUD operations
  */
 public class InvoiceService {
+    
+    /**
+     * Search criteria for invoice filtering
+     */
+    public static class SearchCriteria {
+        private String customerName;
+        private String roomId;
+        private String type;
+        private Date fromDate;
+        private Date toDate;
+        private Double minAmount;
+        private Double maxAmount;
+        
+        public SearchCriteria() {}
+        
+        // Getters and Setters
+        public String getCustomerName() { return customerName; }
+        public void setCustomerName(String customerName) { this.customerName = customerName; }
+        
+        public String getRoomId() { return roomId; }
+        public void setRoomId(String roomId) { this.roomId = roomId; }
+        
+        public String getType() { return type; }
+        public void setType(String type) { this.type = type; }
+        
+        public Date getFromDate() { return fromDate; }
+        public void setFromDate(Date fromDate) { this.fromDate = fromDate; }
+        
+        public Date getToDate() { return toDate; }
+        public void setToDate(Date toDate) { this.toDate = toDate; }
+        
+        public Double getMinAmount() { return minAmount; }
+        public void setMinAmount(Double minAmount) { this.minAmount = minAmount; }
+        
+        public Double getMaxAmount() { return maxAmount; }
+        public void setMaxAmount(Double maxAmount) { this.maxAmount = maxAmount; }
+        
+        @Override
+        public String toString() {
+            return "SearchCriteria{" +
+                    "customerName='" + customerName + '\'' +
+                    ", roomId='" + roomId + '\'' +
+                    ", type='" + type + '\'' +
+                    ", fromDate=" + fromDate +
+                    ", toDate=" + toDate +
+                    ", minAmount=" + minAmount +
+                    ", maxAmount=" + maxAmount +
+                    '}';
+        }
+    }
+    
     private final InvoiceDAOGateway daoGateway;
     
     public InvoiceService(InvoiceDAOGateway daoGateway) {
@@ -22,7 +73,7 @@ public class InvoiceService {
     public boolean createInvoice(InvoiceRequest request) {
         try {
             // Validate request
-            ValidationResult validation = ValidationService.validateInvoiceRequest(request);
+            ValidationService.ValidationResult validation = ValidationService.validateInvoiceRequest(request);
             if (!validation.isValid()) {
                 System.err.println("Validation failed: " + validation.getErrorMessage());
                 return false;
