+++ src/main/java/business/ValidationService.java
@@ -1,10 +1,47 @@
 package business;
 
 /**
- * Validation Service for business rules
+ * Validation Service for business rules with embedded ValidationResult
  * Implements Single Responsibility Principle for validation logic
  */
 public class ValidationService {
+    
+    /**
+     * Result object for validation operations
+     */
+    public static class ValidationResult {
+        private final boolean valid;
+        private final String errorMessage;
+        
+        private ValidationResult(boolean valid, String errorMessage) {
+            this.valid = valid;
+            this.errorMessage = errorMessage;
+        }
+        
+        public static ValidationResult success() {
+            return new ValidationResult(true, null);
+        }
+        
+        public static ValidationResult fail(String errorMessage) {
+            return new ValidationResult(false, errorMessage);
+        }
+        
+        public boolean isValid() {
+            return valid;
+        }
+        
+        public String getErrorMessage() {
+            return errorMessage;
+        }
+        
+        @Override
+        public String toString() {
+            return "ValidationResult{" +
+                    "valid=" + valid +
+                    ", errorMessage='" + errorMessage + '\'' +
+                    '}';
+        }
+    }
     
     /**
      * Validate invoice request
