+++ src/main/resources/presentation/primary.fxml
@@ -22,6 +22,14 @@
     <center>
         <VBox spacing="15" style="-fx-padding: 20;">
             
+            <!-- Search bar -->
+            <HBox spacing="10" alignment="CENTER_LEFT" style="-fx-background-color: #ecf0f1; -fx-padding: 10; -fx-background-radius: 5;">
+                <Label text="Tìm kiếm:" style="-fx-font-weight: bold;" />
+                <TextField fx:id="searchField" promptText="Nhập tên khách hàng hoặc mã phòng..." prefWidth="300" />
+                <Button fx:id="searchButton" text="Tìm kiếm" style="-fx-background-color: #3498db; -fx-text-fill: white;" onAction="#searchInvoices" />
+                <Button fx:id="clearSearchButton" text="Xóa bộ lọc" style="-fx-background-color: #95a5a6; -fx-text-fill: white;" onAction="#clearSearch" />
+            </HBox>
+            
             <!-- Form nhập liệu -->
             <VBox spacing="10" style="-fx-background-color: #f8f9fa; -fx-padding: 20; -fx-background-radius: 5;">
                 <Label text="THÔNG TIN HÓA ĐƠN" style="-fx-font-size: 16px; -fx-font-weight: bold; -fx-text-fill: #2c3e50;" />
@@ -63,7 +71,6 @@
                     <Button fx:id="deleteButton" text="Xóa" style="-fx-background-color: #e74c3c; -fx-text-fill: white; -fx-font-weight: bold;" onAction="#deleteInvoice" />
                     <Button fx:id="clearButton" text="Làm mới" style="-fx-background-color: #95a5a6; -fx-text-fill: white; -fx-font-weight: bold;" onAction="#clearForm" />
                     <Region HBox.hgrow="ALWAYS" />
-                    <Button fx:id="settingsButton" text="Settings" style="-fx-background-color: #9b59b6; -fx-text-fill: white; -fx-font-weight: bold;" onAction="#switchToSecondary" />
                     <Button fx:id="refreshButton" text="Tải lại danh sách" style="-fx-background-color: #3498db; -fx-text-fill: white; -fx-font-weight: bold;" onAction="#refreshTable" />
                 </HBox>
             </VBox>
@@ -91,6 +98,8 @@
                     <Label fx:id="totalInvoicesLabel" text="Tổng số hóa đơn: 0" style="-fx-font-weight: bold;" />
                     <Separator orientation="VERTICAL" />
                     <Label fx:id="totalAmountLabel" text="Tổng doanh thu: 0 VND" style="-fx-font-weight: bold; -fx-text-fill: #27ae60;" />
+                    <Separator orientation="VERTICAL" />
+                    <Button fx:id="monthlyStatsButton" text="Thống kê theo tháng" style="-fx-background-color: #8e44ad; -fx-text-fill: white;" onAction="#showMonthlyStats" />
                 </HBox>
             </VBox>
